% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{can_calculate_ep}
\alias{can_calculate_ep}
\title{Determines whether the exclusion power can be calculated}
\usage{
can_calculate_ep()
}
\value{
TRUE if all the ingredients for calculating EP are present, otherwise
  returns an error message
}
\description{
Determines whether the exclusion power can be calculated
}
