% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_nas_to_string}
\alias{df_nas_to_string}
\title{Substitute NA values for a string in a data frame}
\usage{
df_nas_to_string(df, na_string = "-")
}
\arguments{
\item{df}{a data.frame}

\item{na_string}{the string to put in place of NA values}
}
\value{
a copy of df without NA values
}
\description{
Substitute NA values for a string in a data frame
}
