% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_round}
\alias{df_round}
\title{Round numeric values in a data frame}
\usage{
df_round(df, digits = 3)
}
\arguments{
\item{df}{a data.frame}

\item{digits}{number of digits to round to}
}
\value{
a copy of df with numeric fields rounded accordingly
}
\description{
Round numeric values in a data frame
}
\examples{

df <- data.frame(A = c(0.1234, 0.123))
round_df(df, digits = 2)
}
