% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_candidate_available_ids}
\alias{get_candidate_available_ids}
\title{Calculate IDs that could be available for genotyping}
\usage{
get_candidate_available_ids()
}
\value{
a character vector of IDs that the user may choose to define as
  available for genotyping
}
\description{
An ID can be defined as available for genotyping if: 1. It is present in both
the claim pedigree and the true pedigree, and 2. it has not been already
genotyped.
}
