% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTruthy}
\alias{isTruthy}
\title{Check if something is truthy}
\usage{
isTruthy(x)
}
\arguments{
\item{x}{anything}
}
\value{
a logical
}
\description{
Returns true if the parameter is not false, or similar to false in some way.
}
\details{
Adapted from
https://github.com/rstudio/shiny/blob/5e7447886405392f6397c7e7b8fbb705050ebab4/R/utils.R#L1430
}
