% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{normalize}
\alias{normalize}
\title{Normalize a frequency database}
\usage{
normalize(df)
}
\arguments{
\item{x}{a data frame in allelic ladder format}
}
\value{
a scaled version of the given frequency database
}
\description{
Scales a frequency database so that allele frequencies sum up to one.
}
