% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{read_database_ladder}
\alias{read_database_ladder}
\title{Read allele frequency database in ladder format}
\usage{
read_database_ladder(filename, ...)
}
\arguments{
\item{filename}{a filepath}

\item{...}{further parameters to \code{read.csv}}
}
\value{
a \code{\link{data.frame}}
}
\description{
Read allele frequency database in ladder format
}
